Style the navigation header with flex
Added flex properties to the existing header and ul selectors to
get the header to look like a nav bar. Created two additional divs and
called on the classes in the stylesheet to get the logo and links on
one side and the notification and profile image on the other.